write firebase cloud function for onCreate trigger send email
